# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vijay Transmission Cost Sheet Suite API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change the current user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid new password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Current password incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: List all quotes with optional filters
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: projectRef
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of quotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Save a new quote (auto-assigns revision number)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteInput"
      responses:
        "201":
          description: Quote saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}:
    get:
      operationId: getQuote
      tags: [quotes]
      summary: Get a single quote by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quote details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/by-project:
    get:
      operationId: getQuotesByProject
      tags: [quotes]
      summary: Get all revisions for a client + project
      parameters:
        - name: customerId
          in: query
          required: true
          schema:
            type: integer
        - name: projectRef
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: All revisions for this project
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
  /quotes/{id}/approve:
    post:
      operationId: approveQuote
      tags: [quotes]
      summary: Mark a revision as the vendor-approved quote (clears approval on sibling revisions)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated approved quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
...
  /review/projects:
    get:
      operationId: getProjectsByCustomer
      tags: [quotes]
      summary: List all project refs for a given customer
      parameters:
        - name: customerId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of project refs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 8
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, name, role, isActive]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, user]
        isActive:
          type: boolean
        mustChangePassword:
          type: boolean
        createdAt:
          type: string
...
    Quote:
      type: object
      required: [id, customerId, customerName, projectRef, revision, structureType, kvOption, quotePricePerMt, totalCost, inputs, costBreakdown, generatedByName, legacy, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        projectRef:
          type: string
        revision:
          type: integer
          description: Rev 0 is first quote; increments automatically per client+project
        structureType:
          type: string
        kvOption:
          type: ["string", "null"]
        quotePricePerMt:
          type: number
        totalCost:
          type: number
        steelPrice:
          type: ["number", "null"]
        zincPrice:
          type: ["number", "null"]
        rmPricesId:
          type: ["integer", "null"]
          description: Retained RM price revision used when this quote was calculated.
        rmOffsetsId:
          type: ["integer", "null"]
          description: Retained RM offset revision used when this quote was calculated; null means workbook defaults.
        rmPriceSource:
          type: ["object", "null"]
          description: Dated attribution for the RM price revision used by this quote.
          required: [id, createdAt, createdByName]
          properties:
            id:
              type: integer
            createdAt:
              type: string
            createdByName:
              type: string
        rmOffsetSource:
          type: ["object", "null"]
          description: Dated attribution for the RM offset revision used by this quote.
          required: [id, updatedAt, updatedByName]
          properties:
            id:
              type: integer
            updatedAt:
              type: string
            updatedByName:
              type: string
        inputs:
          type: object
          description: Full JSON of all cost inputs
        costBreakdown:
          type: object
          description: Computed cost breakdown
        generatedByName:
          type: string
        notes:
          type: ["string", "null"]
        approved:
          type: boolean
        approvedAt:
          type: ["string", "null"]
        approvedByName:
          type: ["string", "null"]
        legacy:
          type: boolean
          description: True for quotes computed by the pre-reconciliation engine; shown read-only
        discountMode:
          type: ["string", "null"]
          description: "'pct' or 'abs' — mode of the additional discount applied by salesperson"
        discountValue:
          type: ["number", "null"]
          description: "Discount magnitude: fraction for 'pct' (e.g. 0.05 = 5%) or absolute ₹/MT for 'abs'"
        netQuotePricePerMt:
          type: ["number", "null"]
          description: Quote price after additional discount. NULL when no discount applied.
        createdAt:
          type: string
    QuoteInput:
      type: object
      required: [customerId, customerName, projectRef, structureType, quotePricePerMt, totalCost, inputs, costBreakdown, generatedByName]
      properties:
        customerId:
          type: integer
        customerName:
          type: string
        projectRef:
          type: string
        structureType:
          type: string
        kvOption:
          type: ["string", "null"]
        quotePricePerMt:
          type: number
        totalCost:
          type: number
        steelPrice:
          type: ["number", "null"]
        zincPrice:
          type: ["number", "null"]
        rmPricesId:
          type: ["integer", "null"]
          description: RM price revision loaded by the calculator. The server validates that it exists.
        rmOffsetsId:
          type: ["integer", "null"]
          description: RM offset revision loaded by the calculator, or null when only workbook default offsets applied.
        inputs:
          type: object
        costBreakdown:
          type: object
        generatedByName:
          type: string
        notes:
          type: ["string", "null"]
        discountMode:
          type: ["string", "null"]
        discountValue:
          type: ["number", "null"]
        netQuotePricePerMt:
          type: ["number", "null"]
...